+++ config/params.php
@@ -8,6 +8,8 @@ return [
         'cookieLogin' => [
             'forceAddCookie' => false,
             'duration' => 'P5D', // 5 days, see format on https://www.php.net/manual/dateinterval.construct.php
+            'cookieName' => 'autoLogin',
+            'cookieParams' => [],
         ],
     ],
 ];